+++ python/src/uagents/communication.py
@@ -8,8 +8,8 @@ from typing import NoReturn
 
 import aiohttp
 from pydantic import UUID4, ValidationError
-from uagents_core.crypto import Identity, is_user_address
 from uagents_core.envelope import Envelope
+from uagents_core.identity import Identity, is_user_address
 from uagents_core.models import Model
 from uagents_core.types import DeliveryStatus, MsgStatus
