+++ python/src/uagents/resolver.py
@@ -7,7 +7,8 @@ from datetime import datetime, timezone
 
 import requests
 from dateutil import parser
-from uagents_core.communication import parse_identifier, weighted_random_sample
+from uagents_core.helpers import weighted_random_sample
+from uagents_core.identity import parse_identifier
 
 from uagents.config import (
     ALMANAC_API_URL,
